# ccl: checksum backup/restore files