To make sure the files in the restore are the same ones from the backup
and to make sure no errors occured in any intermediate transmission or
storage.

Seems like the benchmarks are a bit slower for backp, but it's in the
noise.

Closes #13482.

```
name              old time/op    new time/op    delta
ClusterBackup-8     1.09µs ± 8%    1.18µs ±14%    ~             (p=0.095 n=5+5)
ClusterRestore-8    5.25µs ± 7%    4.97µs ± 4%  -5.25%          (p=0.032 n=5+5)

name              old speed      new speed      delta
ClusterBackup-8    113MB/s ± 8%   104MB/s ±13%    ~             (p=0.095 n=5+5)
ClusterRestore-8  23.5MB/s ± 6%  24.7MB/s ± 4%  +5.41%          (p=0.032 n=5+5)

name              old alloc/op   new alloc/op   delta
ClusterBackup-8       188B ±31%      193B ±56%    ~             (p=0.508 n=5+5)
ClusterRestore-8    7.63kB ± 4%    7.86kB ± 5%    ~             (p=0.151 n=5+5)

name              old allocs/op  new allocs/op  delta
ClusterBackup-8       3.00 ± 0%      3.00 ± 0%    ~     (all samples are equal)
ClusterRestore-8      2.00 ± 0%      2.00 ± 0%    ~     (all samples are equal)
```
cc @bdarnell

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/13861)
<!-- Reviewable:end -->