@mjibson 2017-03-01
I think it is worth it to add a test that mutates or appends to a backup to guarantee failure during restore, just to test that this thing is on.

---

Review status: 0 of 7 files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[pkg/ccl/sqlccl/backup.proto, line 35 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/13861#-Ke7aVBvLiWCwA2IJtz-:-Ke7aVBvLiWCwA2IJtz0:bj7pqbe) ([raw file](https://github.com/cockroachdb/cockroach/blob/eec2e496438a78c2f671181d6f7175bdb94ddcce/pkg/ccl/sqlccl/backup.proto#L35)):*
> ```protobuf
>     string path = 2;
>     bytes sha512 = 4;
>     reserved 3;
> ```

The other `reserved` keywords in the proto files have retained their position instead of being sorted to the bottom.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/13861#-:-Ke7aAmQIdDGs3npG5Mk:b8m7v3k)*
<!-- Sent from Reviewable.io -->