@danhhz 2017-03-04
Added a test

---

Review status: 0 of 8 files reviewed at latest revision, 1 unresolved discussion.

---

*[pkg/ccl/sqlccl/backup.proto, line 35 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/13861#-Ke7aVBvLiWCwA2IJtz-:-KeQ6M1Mc4DGi_wLg9qn:bskaa90) ([raw file](https://github.com/cockroachdb/cockroach/blob/eec2e496438a78c2f671181d6f7175bdb94ddcce/pkg/ccl/sqlccl/backup.proto#L35)):*
<details><summary><i>Previously, mjibson (Matt Jibson) wrote…</i></summary><blockquote>

The other `reserved` keywords in the proto files have retained their position instead of being sorted to the bottom.
</blockquote></details>

This isn't quite universal, but done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/13861#-:-KeQBySW4Uw_xWfua4sz:bkdqbhy)*
<!-- Sent from Reviewable.io -->