@mjibson 2017-03-06
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 8 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/13861#-:-Ke_4mkOrTXHCHqKTPrm:bnfp4nl)*
<!-- Sent from Reviewable.io -->