@danhhz 2017-03-06
TFTR!

---

Review status: 0 of 8 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/13861#-:-Ke_5Td2UYeWxEc7xwua:b-814g8k)*
<!-- Sent from Reviewable.io -->